// faber-admin/src/main/frontend-old/src/components/base-tree/interface.ts
namespace BaseTreeProps {
	/**
	 * 运用于BaseTree的node节点
	 */
	export interface TreeNode<T = any, KeyType = number> {
		id?: KeyType;
		parentId?: KeyType;
		name?: string;
		// tree
		label: string;
		value: KeyType;
		isLeaf: boolean;
		children: BaseTreeProps.TreeNode<T>[] | undefined;
		sourceData?: T;
	}

	export interface NodeProps {
		key: number | string;
		title: JSX.Element;
		children: NodeProps[] | undefined;
		name: string;
	}

	export interface ExtraContextMenus {
		key: string;
		menuTitle: string | JSX.Element;
		// TODO 优化e、item的TS类型
		onMenuClick: (e: any, item: any) => void;
	}

	export interface FlatTreeNode {
		key: number | string;
		index: number;
		pid: number | string;
		name: string;
	}
}

export default BaseTreeProps;